requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cpu

altair==4.2.0
    # via streamlit
attrs==22.1.0
    # via jsonschema
blinker==1.5
    # via streamlit
cachetools==5.2.0
    # via streamlit
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via streamlit
commonmark==0.9.1
    # via rich
decorator==5.1.1
    # via validators
entrypoints==0.4
    # via altair
ffmpeg-python==0.2.0
    # via
    #   -r requirements.in
    #   whisper
filelock==3.8.0
    # via
    #   huggingface-hub
    #   transformers
future==0.18.2
    # via ffmpeg-python
gitdb==4.0.10
    # via gitpython
gitpython==3.1.29
    # via streamlit
huggingface-hub==0.11.1
    # via transformers
idna==3.4
    # via requests
importlib-metadata==5.1.0
    # via streamlit
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.17.1
    # via altair
markupsafe==2.1.1
    # via jinja2
more-itertools==9.0.0
    # via whisper
numpy==1.23.5
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
    #   transformers
    #   whisper
packaging==21.3
    # via
    #   huggingface-hub
    #   streamlit
    #   transformers
pandas==1.5.2
    # via
    #   altair
    #   streamlit
pillow==9.3.0
    # via streamlit
protobuf==3.20.3
    # via streamlit
pyarrow==10.0.1
    # via streamlit
pydeck==0.8.0
    # via streamlit
pygments==2.13.0
    # via rich
pympler==1.0.1
    # via streamlit
pyparsing==3.0.9
    # via packaging
pyrsistent==0.19.2
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   pandas
    #   streamlit
python-multipart==0.0.5
    # via -r requirements.in
pytz==2022.6
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
regex==2022.10.31
    # via transformers
requests==2.28.1
    # via
    #   huggingface-hub
    #   streamlit
    #   transformers
rich==12.6.0
    # via streamlit
semver==2.13.0
    # via streamlit
six==1.16.0
    # via
    #   python-dateutil
    #   python-multipart
smmap==5.0.0
    # via gitdb
streamlit==1.15.1
    # via -r requirements.in
tokenizers==0.13.2
    # via transformers
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
torch==1.13.0
    # via
    #   -r requirements.in
    #   torchaudio
    #   whisper
torchaudio==0.13.0
    # via -r requirements.in
tornado==6.2
    # via streamlit
tqdm==4.64.1
    # via
    #   huggingface-hub
    #   transformers
    #   whisper
transformers==4.24.0
    # via whisper
typing-extensions==4.4.0
    # via
    #   huggingface-hub
    #   streamlit
    #   torch
tzdata==2022.6
    # via pytz-deprecation-shim
tzlocal==4.2
    # via streamlit
unidecode==1.3.4
    # via -r requirements.in
urllib3==1.26.13
    # via requests
validators==0.20.0
    # via streamlit
watchdog==2.1.9
    # via streamlit
whisper @ file:///app/whisper
    # via -r requirements.in
zipp==3.11.0
    # via importlib-metadata
